Fix the generated tarball on Mac OS X
We were seeing an error like

   chapel-1.28.0/compiler/dyno/tools/chpldoc/BUILD_VERSION: Skipping hardlink pointing to itself: chapel-1.28.0/compiler/dyno/tools/chpldoc/BUILD_VERSION

It turns out that compiler/dyno/tools/chpldoc/BUILD_VERSION was a
symbolic link that we listed twice in the tar command so it was included
twice in the tar archive.

Changing the gen_release script to include that file only once in the
archive solves the problem.

---
Signed-off-by: Michael Ferguson <mppf@users.noreply.github.com>