Don't add (dyno-based) chpldoc sources to tags/ebrowse output
The ALL_SRCS Makefile variable is used to determine which sources
should have tags and/or ebrowse information generated for them.
Traditionally, the chpldoc sources and the `chpl` sources have been
one and the same, but with the recent dyno-based rewrite of chpldoc,
this is no longer the case.  For that reason, it doesn't really make
sense to include its sources in the tags/ebrowse output since it's a
distinct binary / code base.  As a result, this PR removes its sources
from the ALL_SRCS variables.  If/when we want tags/ebrowse information
for chpldoc itself, we can do that as a distinct tags/ebrowse
artifact.

---
Signed-off-by: Brad Chamberlain <bradcray@users.noreply.github.com>