Remove deprecated versions of some IO functions
Removes the deprecation warnings for start/end arguments and the iohint
argument.  All of these warnings were added last release and have been
cluttering the file.  We want to make a change to the region argument behavior,
to simplify that change remove the old versions instead of leaving them lying
around.

Impacts:
- channel.seek
- openreader
- file.reader
- file.lines
- file.writer

----
Signed-off-by: Lydia Duncan <lydia-duncan@users.noreply.github.com>