Propagate notes from new error system to production compiler
This is done by expanding on the "dummy ErrorWriter" approach; an
ErrorWriter is created that tracks the output of notes and other content
from the error. The information gathered by this new ErrorWriter is then
used to create a matching ErrorMessage that can be interpreted by the
production compiler in the already-established way.

Signed-off-by: Danila Fedorin <daniel.fedorin@hpe.com>