Fix sporadic .bad mismatch in nestedOn
This test:

    test/optimizations/remoteValueForwarding/gmp/nestedOn.chpl

has nested on-stmts and the inner one halts. We sporadically
get warnings like: `GASNET WARNING(Node 0): int AMUDP_DrainNetwork(ep_t)
returning an error code: AM_ERR_RESOURCE` for this test.

Just add a prediff to filter out any extra messages and only
look for our expected error.

Once the .future is resolved, we will be getting mismatches on both .bad
and .good. At this point the prediff needs to be removed.

This is similar to #12937 and #14837 for these tests:

    test/multilocale/bradc/needMultiLocales/writeThisUsingOn.prediff
    test/multilocale/local/diten/test_local2.prediff

The solution idea was proposed by @ronawho.

Signed-off-by: Vassily Litvinov <vasslitvinov@users.noreply.github.com>