Make openUrlReader use the new file.reader behavior without flags
Adds a new argument to file.readerHelper to indicate the call is coming from
openUrlReader.  Make the new code get called with that flag, regardless of if
`useNewFileReaderRegionBounds` is called, because it's an implementation detail
for URL.chpl and we don't want to accidentally force the user into the new
behavior for their calls.

----
Signed-off-by: Lydia Duncan <lydia-duncan@users.noreply.github.com>