Addressed reviewer's comments.
CHPL_QTHREAD_TOPOLOGY is a build-time environment variable not available
at run-time. The qthreads Makefile now sets a compiler macro based on its
value, and this macro is used to initialize a variable that can be checked
at run-time.

Added a comment as to why the CPU ID string is 5 times the number of CPUs.

Moved QT_CPUBIND code to setupAffinity();

Signed-off-by: John H. Hartman <jhh67@users.noreply.github.com>