Remove old support for tracking release-over-release performance
In the past, we would manually run the current benchmarks with older
chapel releases and graph the results. This was useful for weeding out
changes to benchmarks and machines to give us a simpler view of how
performance of the compiler/modules/runtime changed over time.

However, we haven't really done this since ~2018. This information was
really useful during our "5 year push" to improve performance, but
performance is pretty stable nowadays and we track weekly performance
triage much more closely so it's rare that any regressions slip in.

Given that we haven't used this information in years, I think we can
remove the infrastructure that we used to collect the results and the
infrastructure we use to overlay the release-over-release results with
nightly.

Signed-off-by: Elliot Ronaghan <ronawho@gmail.com>