Type comparison operators are unstable, not deprecated
Per discussion in #20666.  These are attached to docs-only functions, so the impact is docs-only.
Since they're seen only in docs, the doc's preceding specification for e.g.
>= returns isSubtype(b,a)
means the unstable warning can just refer to isSubtype without having to spell out the order of arguments.

Also update the methods' comments not to recommend the unstable
operators.

---
Signed-off-by: Paul Cassella <gitgitgit@manetheren.bigw.org>