Disable ugni MR checks when the cache is enabled
Disable memory registration extent checks when the remote cache is
enabled since we've been getting some false-positives. These checks were
added in 19127 and were meant to help catch subtle implementation bugs
where a single RDMA transfer crossed multiple memory registration
regions. Our MR extents are precise (in that they track the allocation
start and end exactly) but the cache can read to the end of a page,
which caused false-positives. We tried to quiet them in 19516 and 20642,
but we've still been seeing errors sporadically in nightly. Reading to
the end of a page is safe since the kernel's memory registration covers
that so it was really just our checks being overly conservative. We
could probably tighten up our checks, but given that ugni is no longer
under active development I think it's ok to just disable these checks
when the cache is on.

Resolves Cray/chapel-private 3341

Signed-off-by: Elliot Ronaghan <ronawho@gmail.com>