Track declaration visibility modifiers in Scope.
To preserve the ability to "borrow" an owned list of IDs in a scope
(which is efficient if the scope has a _lot_ of IDs with a particular
name, such as many overloads of a particular function), this commit
makes the BorrowedIds track whether or not visible symbols are to be
reported. A new non-trivial iterator is also added for BorrowedIds,
which skips over entries in the scope when private symbols are being
excluded, and doesn't when private symbols are requested to be left in.
Thus, the iterator returned from a BorrowedIds will contain the correct
symbols without the need to create new (filtered) lists.

This commit also modifies the symbols stored in Owned and Borrowed lists
to include the symbols' visibilities, which is necessary to determine
whether or not they should be reported.

An alternative was to split the OwnedIds into multiple lists, but this would
mean either duplication (one list containing all IDs, and one list containing
IDs only visible publically) or losing order (public IDs, then private
IDs). This approach would also require a non-trivial iterator.

Signed-off-by: Danila Fedorin <daniel.fedorin@hpe.com>