Treat `export` functions as potentially remote
Resolves #20861

When widening pointers/references, include `export` functions
in the set of functions that could be executed on a non-zero locale
-- in the case when an `extern` function could be executed remotely.
This is because an `extern` function can potentially execute any `export`.

When we did not do it, the struct type of `globalArr` in the generated code
had the field `_instance` as a non-wide pointer, see below. So whenever
it was accessed from a non-0 locale, the generated code refereced local memory
using a memory address from Locale 0.
```c
typedef struct chpl__array_DefaultRectangularArr_1_int64_t_F__real64_int64_t_s {
  int64_t _pid;
  DefaultRectangularArr_1_int64_t_F__real64_int64_t _instance;
  chpl_bool _unowned;
} _array_DefaultRectangularArr_1_int64_t_F__real64_int64_t;
```

The `export` functions in internal modules are excluded from consideration.
This is our of precaution to avoid widening of some globals that do not
needed it and so hurting performance unnecessarily.

Signed-off-by: Vassily Litvinov <vasslitvinov@users.noreply.github.com>