Allow per test-run customized environments
Previously, all runs of the same test (caused by having multiple
compiler and/or execution options) would use the same execution
environment, generated from the EXECENV and .execenv files prior to the
first test run. This made it impossible for each test run to use a
different environment. With this change, the EXECENV and .execenv files
are read on each run, and both are passed the compiler options and
execution options as arguments, allowing them to specialize the
environment for the test run.

Signed-off-by: John H. Hartman <jhh67@users.noreply.github.com>