Update a .good file broken by d41692c7557fae36a72397a414ff89886f9e4818
In making a last-minute update to a test in my PR, I failed to take
into account that the test file might be used from other directories,
which caused this test to break.  This should fix it.  I did not find
other cases that were broken due to this with an additional check.

---
Signed-off-by: Brad Chamberlain <bradcray@users.noreply.github.com>