Improve printing of atomic types
Print atomic types as they appear in source code and not using the
internal types. For example print `atomic int(64)` instead of the
internal `AtomicT(int(64))` name.

This is a nicer way to print and enables removing some network-atomic
specific good files since processor atomic (`Atomic*`) and network
atomic (`RAtomic*`) types are no longer printed differently.

This is a logical follow on to 13842, though a slight difference
compared to `sync` vars is that we have a different a non-generic type
for atomic bools so we need a different renaming strategy there. Longer
term we expect dyno to eliminate the need for all this renaming, so the
duplication in the short term doesn't seem too bad.

Signed-off-by: Elliot Ronaghan <ronawho@gmail.com>