Fix sporadic failure in remote-array-write-access
This should fix the sporadic failure in `performance/ferguson/remote-array-write-access.chpl` when compiled with `--cache-remote` enabled.

The failure mode, most recently observed in our `gasnet.darwin-m1` configuration, is:
```
ERROR -- TOO MANY GETs, 2 EXCEEDS MAX OF 1
```

When Locale 1 issues a cache-get of A._instance, which is a wide pointer, it is typically done with a single get_nb. However, sometimes the _instance field is split across two cache pages, so one cache-get results in two get_nb communications. This PR allows this test to accept either 1 or 2 get_nb.

The solution idea was proposed by @ronawho.

Signed-off-by: Vassily Litvinov <vasslitvinov@users.noreply.github.com>