Add a mechanism to disable comm diags for a specific task
Previously, comm diagnostics could only be disabled at a locale
granularity, but there are some cases where we just want to disable comm
for a specific function call for the current task. This adds a mechanism
that uses task-locale-storage to temporarily disable comms for the
current task and replaces the previous locale-wide disabling code.

This is motivated by some upcoming work to not track communication from
end counts, but was useful for replacing an existing mechanism so I
wanted to split it into a separate PR.

Note that I removed the comm diags disabling from chapel-gpu-diags since
that isn't broadcasting code to enable/disable diags so I don't think
it's required and we don't disable for similar broadcasting in memory
diagnostics code. If we decide we should disable when broadcasting
everywhere we could do that directly in `chpl_comm_bcast_rt_private`.

In the future it may also be interesting to consider enabling comm diags
for a specific task only as there have been times I just want to track
only some comm from a locale. I didn't do that here in part because I
wasn't ready to do the API design, but also because that would require
getting TLS for every comm diag call and that has a non-zero cost.

Signed-off-by: Elliot Ronaghan <ronawho@gmail.com>