Avoid a segfault after hitting a USR_FATAL_CONT
The code reorg I did in #20891 led to a segfault for CHPL_COMM!=none
compilations that hit a USR_FATAL_CONT() within the compiler in the
event that we had a 'yield' within an 'on'-clause.  The cause was that
the 'testBlock' was never created, but the downstream code assumed it
was, so dereferenced a NULL pointer.  Here, I took the approach of
removing the 'break;' after the error message is printed in order to
build the 'testBlock' as usual.  This had the effect of causing the
error messages to print out twice per line, but this is appropriate
because there are two iterators containing the problem being invoked
on each line (when there's only one, just one error is printed).

---
Signed-off-by: Brad Chamberlain <bradcray@users.noreply.github.com>