Revert "Merge pull request #20880 from arezaii/chpldoc-install-fix"
This reverts a fix for installing chpldoc that exposed a shortcoming
in chpldoc where it doesn't do a search for possible CHPL_HOME values
if CHPL_HOME is not detected in the environment or passed using the
--home option.
Once the fixes have been made for chpldoc, we will un-revert this
revert.

This reverts commit 85c7720ea56ea8762d4b4b4718c89214427ab2b2, reversing
changes made to 86393aed5d972fa50e737c5807a2b9471fc8fffc.

Signed-off-by: arezaii <ahmad.rezaii@hpe.com>