Make these more official deprecations
These were added prior to the `deprecated` keyword, so used the old strategy of
inserting a compilerWarning into the body instead of relying on the deprecated
keyword.  While this worked reasonably for the uses that were tested, there were
cases where this meant we were not reporting where the call was coming from,
making it harder to resolve.  This change adjusts them to use the keyword, to
help me track down where the call is coming from and avoid this happening for
other cases.

----
Signed-off-by: Lydia Duncan <lydia-duncan@users.noreply.github.com>