Use pkg-config for zmq package tests
Previously, the ZMQ package tests required a system install of libzmq to
work. Our skipif would look for a libzmq library and if it was found
assume that a zmq.h header would be found too. However, this isn't
always the case and we have systems that have the libzmq library but not
the headers, which caused the tests to run and then fail. To improve
this, use pkg-config to check if zmq is installed and to get appropriate
compilation cflags/libs.

We've run into the zmq library only issue before and added a workaround
in 20191 that let us manually skip the tests, but this makes that check
automatic, at least for the package tests. We'll want to do something
similar for the multi-locale interop tests at some point too, but I
think that will require updating the compiler or makefile makefile
support for mli to use pkg-config to get cflags/libs.)

Tested `start_test --respect-skipifs test/library/packages/ZMQ/` on
 - [x] machine with system libzmq.so and zmq.h (passed)
 - [x] machine with spack installed zmq (passed)
 - [x] machine with brew installed zmq (passed)
 - [x] machine with only libzmq.so and not zmq.h (skipped)

Motivated by Cray/chapel-private 4011

Signed-off-by: Elliot Ronaghan <ronawho@gmail.com>