Skip `--static` test if libpthread.a isn't found
Currently, the `--static` compiler flags maps to a backend flag like
`-static`, which tries to create a fully static executable. This is
support by fewer and fewer distributions anymore since they don't ship
static versions of libraries we need like pthreads. We could probably
stand to revisit what `--static` means or whether we want to continue to
have the flag, but for now I just want to prevent the test from failing
on systems that lack static versions of libraries we need.

Part of Cray/chapel-private 4011

Signed-off-by: Elliot Ronaghan <ronawho@gmail.com>