ofi NIC selection when there is one locale per socket
Adds functionality the ofi layer to assign NICs to locales when there is one
locale per socket on a multi-socket node. This is done deterministically, so
that locales do not need to communicate to decide which NIC to use. There are
several complications, the first being that libfabric returns providers in
priority order, with the highest performance provider first. Providers with
equivalent performance are not returned in any particular order. When there are
multiple locales running, each in its own socket, each locale should use the
NIC associated with its socket, if one exists. The function
chpl_topo_selectNicByType is called by the ofi layer to find the best NIC to
use of the same type as the specified NIC, so that if the first provider is in
the wrong socket the ofi layer will iterate through providers until it finds
one in the correct socket.

Another complication is that hwloc topology layer and the libfabric layer have
different naming conventions for NICs. During the selection process a NIC is
uniquely identified by its PCI address.

NIC selection works correctly if some sockets do not have a NIC or if some
sockets have more than one NIC. Locales without a NIC in their socket will
use "extra" NICs if some sockets have more than one and will double-up on the
assigned NICs otherwise. If there are more sockets than NICs then some NICs
will go unused (a single locale only uses a single NIC).

This functionality currently does not support Infiniband or any other NIC that
is reported by hwloc as being of type HWLOC_OBJ_OSDEV_OPENFABRICS.

NIC selection requires that the hwloc layer return PCI devices so the
--disable-pci flag is removed from hwloc configuration. This may cause
hwloc to depend on the pciaccess library if it is available.

Signed-off-by: John H. Hartman <jhh67@users.noreply.github.com>