Account for ref+out intents with virtual methods
Resolves #21073

This fixes a bug where an actual argument was always considered
a "use" upon a call to a virtually-dispatched method even when
the corresponding formal had a ref or [in]out intent.

The bug was observed when such an actual argument was an array
indexing expression, which involved return intent overloading.
Because of the bug, the return-value overload was chosen
instead of the return-ref one.

Signed-off-by: Vassily Litvinov <vasslitvinov@users.noreply.github.com>