Add documentation about what errors can be thrown to `open`
This function was only documented as throwing `SystemError`s, however, it could
throw more specific subclasses of `SystemError` and if a user caught those
errors, they might be able to adjust their program in some way to allow it to
keep running.  Document those subclasses so that a user knows what they could
catch.

----
Signed-off-by: Lydia Duncan <lydia-duncan@users.noreply.github.com>