Avoid flushing aggregators in lockstep
When flushing, previously all aggregators would flush to locale 0, then
1 and so on in psdueo lockstep. This created a many-to-one bottleneck,
especially at scale. This updates aggregators to start flushing to the
locale they last aggregated a value for. Keeping track of the last
locale is trivially fast and getting rid of the lockstep comm behavior
improves performance, particularly at scale.

This change was made to Arkouda's aggregators in Bears-R-Us/arkouda 1826
and this PR effectively pulls in an "upstream" change. Eventually we
just want Arkouda to use Chapel's aggregators, but there have been
frequent enough changes to aggregators that it's been hard to time.

Signed-off-by: Elliot Ronaghan <ronawho@gmail.com>