Dial down rationale for deprecation
Damian reminded me that we have some open issues about interpreting
% and %= on reals as fmod() calls, so toning down my original
rationale for the deprecation.  While we talked about not deprecating
things without a proposed replacement, in this case the current
feature is so broken, I don't think it applies (in fact, we could
arguably just remove it without deprecation, but since Arkouda
proved to be using it (ish) with 1.28, I don't want to leave anyone
hanging or surprised.

---
Signed-off-by: Brad Chamberlain <bradcray@users.noreply.github.com>