Fix some tests that broke due to the FileSystem permissions deprecation
I added a local version of 'copyMode' that uses local versions of 'getMode' and
'chmod' to the FileSystem.copy function. It would be nice to make 'copy' work
with only the OS.POSIX functions, but in at least one case it isn't currently
working that way.

Updated tests to work with FileSystem permissions functions deprecated.

Signed-off-by: David Iten <daviditen@users.noreply.github.com>