Update the tmbundle update instructions for testing highlight changes.
Github Lightshow appears to now be deprecated or paid-only, so replace the
instructions involving it with ones that worked this time around.

While here, also mentioned that you need to merge a PR with the changes, since
the old instructions could be construed to imply that merely making the changes
on your local copy was sufficient.

----
Signed-off-by: Lydia Duncan <lydia-duncan@users.noreply.github.com>