Fix some tests with call init deinit changes
Most notable change here is to infer the types of
generic 'out' formal arguments from the function body.
This happens in resolveFnCall so that the rest of resolution
can work with the real 'out' formal types (similarly
to how the return type is also inferred at this point).
The main advantage of doing it here (and not earlier, say
in instantiating signatures) is that candidate selection can
happen without resolving the bodies.

---
Signed-off-by: Michael Ferguson <mppf@users.noreply.github.com>