Don't execute slow running 2021 adventOfCode programs
adventOfCode/2021 is our slowest directory when paratesting due to the
number of tests in the dir and long execution time of a few tests. To
improve this, just skip executing tests that took more than 30s to run
since we don't want to spend time optimizing the codes.

This includes `noexec`ing day23a/day23 (the longest running at 280s and
115s respectively) and day21a which also took 30s even with `--fast`.
For day{4,12,15} just skip the slower execopts. This gets rid of all
long running executions and drops the total testing time for the
directory from 700s to 300s.

Part of Cray/chapel-private 3211

Signed-off-by: Elliot Ronaghan <ronawho@gmail.com>