Stop overriding user supplied `CHPL_*_MEM`
Our chplenv scripts typically just define defaults, but allow users to
override even if they select unsupported configs. However, `CHPL_MEM`
was forcing the default for some cases and wouldn't allow users to
override. This wasn't a huge issue because jemalloc doesn't work on
cygwin or as a target allocator for ARM macs. However, it also prevented
using jemalloc as a host compiler on macs, but that configuration should
work, we just haven't figured out rpath stuff to make it the default.

Part of Cray/chapel-private 4113

Signed-off-by: Elliot Ronaghan <ronawho@gmail.com>