NOTEST studies/ssca2/rachels
This test has some of our slowest compilation times. While looking for
ways to speed it up, I noticed that it was originally a clone of the
main ssca2 with a different barrier implementation added in b32f7a4eb5.
That barrier was then folded into the main ssca2 in 82837cc656, so I
don't think there's a reason to test the clone anymore.

This version and main have drifted apart over time as main has been
updated, but the diff doesn't look interesting enough to keep both so
here I'm notesting it.

If anybody thinks it's worth keeping, I'd lean towards dropping `--fast`
from the compopts and maybe dropping a few of the 5 compopts.

Part of Cray/chapel-private 3211

Signed-off-by: Elliot Ronaghan <ronawho@gmail.com>