Handle --mem correctly in slurm-srun launcher tests
The testVFlag and testVerboseFlag tests assumed that --mem=0 is passed
to srun. However, the CHPL_LAUNCHER_MEM variable controls the --mem
argument and can break these tests. If CHPL_LAUNCHER_MEM is not set
then the default value of 0 is used with --mem, otherwise if
CHPL_LAUNCHER_MEM is set to "unset" the --mem option is not passed to
srun, otherwise the value of CHPL_LAUNCHER_MEM is used with --mem. The
exact value passed to --mem is not tested, just whether or not it
appears as an argument to srun.

Signed-off-by: John H. Hartman <jhh67@users.noreply.github.com>