Split performance/ferguson into subdirs
This dir used to take 25 minutes to paratest and now it's down to 7.5
for the tuple dir and sub 5 for everything else. This removes a rather
long tail from gasnet paratesting.

If needed, we could further subdivide the tuple dir into tuple and
tuple-small or something (or just get rid of the tuple small version.)

File movement done with:

```sh

for type in tuple record class array; do
    gsed -i "s:remote-$type:$type/remote-$type:g" $CHPL_HOME/test/CC-GRAPHFILES
    rm remote-$type*.{cc-compopts,cc-keys,graph}
    mkdir $type
    mv remote-$type* $type/
done

./gengraphs.bash
```

Part of Cray/chapel-private 3211

Signed-off-by: Elliot Ronaghan <ronawho@gmail.com>