Snapshot of progress on tracking initializer phases during error handling
Most of the changes are to InitErrorHandling - at this phase in compilation, it
can no longer rely on the AST structure that initializerRules was relying on to
detect `this.init`, `super.init` and `this.complete`, so it needed to implement
its own version of those.

For now, have an instance of the state machine saved in the ErrorHandlingVisitor
but may need some adjustment due to the need for special case detection in
conditionals and loops.

----
Signed-off-by: Lydia Duncan <lydia-duncan@users.noreply.github.com>