Paratest the slowest directories first
Paratest farms out directories to worker nodes to parallelize and speed
up testing. Previously, it farmed out in lexicographic order, but this
could result in a long tail of tests if a long running directory was
handed out near the end. This updates paratest to try and run the
slowest directories first. It does this by extracting directory timings
from a previous test log into a file in `$HOME/.chpl/paratest_times/`.

For this first version, only save the historical data if we're running
the full test suite and there were less than 10 failures. Results are
saved into comm layer specific files since `comm==none` vs `comm!=none`
will likely have a big impact on what tests are run and how long they
take.  Longer term, we may want to include other things that impact
results (valgrind/asan, headnode/slurm-partition) and we'll want to be
able to store and use results if a subset of tests are run. Using old
results makes sense, but it's not obvious when we should override
historical results (don't want to throw away a full test suite of data
when running a few directories.)

This should have a non-trivial impact on parallel testing speed by
getting rid of a long tail. On chapdl this takes a std paratest from 13
minutes to 8 minutes.

The script to sort directories also has code to identify long running
tests/directories. I've been using an earlier version of this to find
and speed up slow running executables and dirs and thought it'd be
useful to have committed.

Part of Cray/chapel-private 3211

Signed-off-by: Elliot Ronaghan <ronawho@gmail.com>