Increase the amount of oversubscription for paratesting
Bump number of paratest clients from ~4 to 6. Historically, there was no
advantage to running more clients since we got stuck waiting for a long
tail, but now that we're running slower dirs first there is a speedup
and this should shave a couple minutes off a paratest.

We also used to have more complicated logic to drop the number of
clients when doing a gasnet run or if another paratest was running to
try to limit interference. In practice we haven't had any issues with
interference so I want to try a fixed number of clients.

Part of Cray/chapel-private 3211

Signed-off-by: Elliot Ronaghan <ronawho@gmail.com>