Fix issue with genGraphs when .dat files are missing
Previously, when only a subset of the tests for a `.graph` file were
run, genGraphs would end up using the first `n` graphkeys instead of
using the graphkeys associated with the .dat file. So if you just did
something like:

```
start_test --performance  test/studies/shootout/nbody/bradc/nbody-blc-ref.chpl
```

You'd end up with the first graphkey of `original` instead of the
correct `ref var version`. This shouldn't be an issue for nightly
testing since we should be generating values for all .dat files in any
given graph, but it's still concerning. Attempt to fix that here by
just not adding graphkeys for datfiles that don't have data.

The alternative would be to fill in the columns later in when we're
looking for values. The tradeoff is whether we skip generating graphs
and legend entries for graphs/keys without values or we generate empty
ones.

It's been a while since I've touched genGraphs and it's a mess so I'm
not sure this is correct yet. To test I think I'd want to generate
graphs before and after this change on chapcs nightly data and verify
the graphs are identical. As a reminder for my future self, to manually
run genGraphs it should be something like:

```
./util/test/genGraphs.py -p $CHPL_TEST_PERF_DIR -o $CHPL_TEST_PERF_DIR/html -t test -r avg -g test/GRAPHFILES -n chapcs
```

Signed-off-by: Elliot Ronaghan <ronawho@gmail.com>