Fix a problem in populateEnvMap
populateEnvMap used to enumerate key/value pairs in the order that they
were emitted by printchplenv. However, PR #20464 changed it to use a
helper routine that returns an unordered_map. This caused a problem
because populateEnvMap was relying on CHPL_TARGET_COMPILER being
processed before CHPL_TARGET_CPU (since useDefaultEnv uses
CHPL_TARGET_COMPILER to decide what to do for CHPL_TARGET_CPU).

This commit fixes it by separately computing a bool to indicate if we
are working with a Cray PrgEnv and passing that bool to useDefaultEnv.

---
Signed-off-by: Michael Ferguson <mppf@users.noreply.github.com>