Start storing class structure type instead of pointer type for TypeSymbols
With opaque pointers we now need to store store the class structure type
instead of the class pointer type because we need this information in a few
places. So when storing this type don't make it a pointer.

Change the name of 'TypeSymbol::llvmType' to 'TypeSymbol::llvmImplType' and
add accessors to get the field either as a pointer or as the underlying
structure. Update accesses of llvmType to use the accessor that returns
a pointer for classes.

The accessor that returns the structure is currently unused so is commented
out to avoid unused function warnings. It will be used shortly.

Signed-off-by: David Iten <daviditen@users.noreply.github.com>