Preserve `this.complete` and `complete` calls in initializers until errhandling
While `setcid` could be used to detect when the type was done with Phase 2 in
classes, it doesn't exist for records.  Which meant we needed to preserve the
presence of those indicators until error handling in order to do proper
detection.  Decided to create a new primitive and replace the detected AST at
normalize with that primitive.  This greatly simplified the check at error
handling.  Since we're done with it after that point, remove it from the
function to avoid problems with the primitive later in compilation.  Also,
clean up the state tracker at that point, otherwise it won't get cleaned up.

----
Signed-off-by: Lydia Duncan <lydia-duncan@users.noreply.github.com>