Fix handling of conditionals
Two fixes:
- Was accidentally overriding the old state instead of saving it while
traversing the "then" branch.  Made new state trackers for the then and else
branches and used them, merging their results into the original and cleaning
them up.
- Was accidentally failing to go into conditionals when removing `this.complete`
when we're done checking the error handling of the function.  Replaced the old
traversal with a new one that gathers all the call expressions ahead of time
so I didn't have to write the traversal appropriately for all nodes myself.

With this change, I think we're finally handling phase tracking appropriately

----
Signed-off-by: Lydia Duncan <lydia-duncan@users.noreply.github.com>