Add checking to ensure we don't allow calls to throwing funcs before phase 2
I think this is working, but there's probably some testing updates I still need
to do.

Basically, allows all throwing functions during phase 2, regardless of if their
result is caught.  If a call to a throwing function is in phase 0 or phase 1, it
needs to be in a try! without a catch clause or it will also trigger an error.

----
Signed-off-by: Lydia Duncan <lydia-duncan@users.noreply.github.com>