Add working, and non-working, tests of slicing with strided unbounded ranges
My PR makes the futures on bool and enum work differently than they
would on master today because I compute an alignment for those ranges
when applying 'by' to them so that I can iterate over them.  Actually,
main computes an alignment on the ranges as well, but then somehow
ignores that alignment when slicing, though I'm not sure how...

I think the right thing to do here might be to store '-1' as a
sentinel value for unaligned/ambiguous ranges in chpl_by_help and then
to have the slicing/iterator code pick up on this and adjust to the
array bounds / compute the alignment implied by the enum/bool range,
respectively?

---
Signed-off-by: Brad Chamberlain <bradcray@users.noreply.github.com>