Add future for #21193
Add a future for #21193

I am adding .execopts with --memLeaks because I suspect that even if
the test compiles, it will leak an object or two. See a comment in
the source.

Once this test works without memleaks, the --memLeaks should be removed
so that most testing configs run without it just like for most other tests.

Signed-off-by: Vassily Litvinov <vasslitvinov@users.noreply.github.com>