Retire reduce-with-for-in-par-iter.reduce2.chpl
This .reduce2.chpl test used to be a future for one of the cases in

    reduce-with-for-in-par-iter.chpl
    reduce-with-forall-in-par-iter.chpl

This future was resolved as of #12131. This PR
uncomments that case in the the first of the two tests above
and removes .reduce2.chpl.

This case has already been uncommented in the second
of the two tests above, so no changes are needed there.

This PR supersedes #20724 by @cassella,
who drew my attention to this change.

Signed-off-by: Vassily Litvinov <vasslitvinov@users.noreply.github.com>