Add a chplType field to the LVT
Add a 'Type* chplType' field to the LVT to track pointer types. Adjust the
LayeredValueTable::addGlobalValue() function to take a Type* argument and
set the chplType field.

Adjust all calls to addGlobalValue() to pass a Chapel Type* to the new
argument.

Adjust LayeredValueTable::getValue() to set the returned GenRet's chplType
to the field set by addGlobalValue().

Signed-off-by: David Iten <daviditen@users.noreply.github.com>