Simplify a range iterator's check for an empty/degenerate range
In implementing #21332, I changed a degenerate range check from
comparing ._low and ._high to something involving more helper routines
and translating between representations.  This resulted in generating
code that caused the C compiler to warn about comparisons that were
tautologically true or false (like how the following will never be
true):

```
if ((x == y) > true)
```

Here, I'm changing them back, which simplifies the code for readers,
doesn't break any of the new tests I added, and makes the warnings go
away (ironically, probably just by making the values less evident to
the C compiler :).

We also saw some performance regressions in our nightly runs, and I'm
wondering whether this could have been a factor.  On one hand, it's a
long-shot that this happens to be it; on the other hand, it did
complicate an expression somewhat.

---
Signed-off-by: Brad Chamberlain <bradcray@users.noreply.github.com>